add circles, avg axis and css
add circles append. The x value corresponds to d.precio, the y value to
d.valor and the r value to d.volumen.
Add avg axis that change depending on the values of the csv. (d.precio
and d.valor)
add a main.css for styling. avg axis are drawn with a light alpha.